janus-lens can now issue commands to janus-dossier
Implemented a way for janus-lens to send commands to janus-dossier in the form of a text file containing the command. janus-dossier then processes the command as if it was a user command using the janus/commands handlers.
Added functionality to the plus and minus buttons to allow for increasing/decreasing requestedQuantity by 1.
Added functionality to select multiple items from the list and issue pause/remove commands for the selected items.
Removed some unnecessary code and renamed some functions to make more sense.