Make UI responsive
Janus-lens now returns from `updateInfo()` immediately if it has been less than 3 seconds since the last update.

This allows us to dispense with the `sleep(2)` in the main loop, which means buttons now respond immediately (probably) and on every press (probably), though it is still a bit tricky to actually hit the smaller ones.

The 3 second interval can be changed as appropriate. It is set to 3 right now to make it easier to read status/debug messages from `setStatus()`. If it is set to less, the status messages are likely to disappear before we can read them. I plan to borrow an idea from Boeing, namely the Master Caution system, to fix this permanently.

Additionally, `btnClose` now has its `action` set in the proper way.