Made status messages customizable
janus-dossier now sets a simple "status" like "craftable", "paused" etc. janus-lens has a table of all these statuses to which we assign a flavour text and color. This text and colour are then used for the UI.

Another change is the addition of performCalculation() function to janus-lens. This should help with the modify commands in the future. For now the minus and plus buttons have been changed to increase or decrease the requestedQuantity to the closest "computer number"